Restructure the Diabetic Disease Folder
