+++ examples/healthcare/application/Diabetic_Disease/Diabetic_Readmission_Prediction/README.md
@@ -41,5 +41,5 @@ To address this issue, we use Singa to implement a machine learning model for pr
 
 ## Command
 ```bash
-python train_mlp.py mlp diabetic
+python train.py mlp diabetic
 ```
