+++ examples/healthcare/application/Diabetic_Disease/Diabetic_Retinopathy_Classification/run.sh
